revert: remove `getByDocId` optional parameter
This reverts commit 56f7e18ae618e6d36dcfdf478ffe42f879aecd2e and
replaces it with an internal-only function, `getByDocIdIfExists`.
