+++ src/datatype/index.d.ts
@@ -18,7 +18,7 @@ import TranslationApi from '../translation-api.js'
 type MapeoDocTableName = `${MapeoDoc['schemaName']}Table`
 type GetMapeoDocTables<T> = T[keyof T & MapeoDocTableName]
 /** Union of Drizzle schema tables that correspond to MapeoDoc types (e.g. excluding backlink tables and other utility tables) */
-type MapeoDocTables =
+export type MapeoDocTables =
   | GetMapeoDocTables<typeof import('../schema/project.js')>
   | GetMapeoDocTables<typeof import('../schema/client.js')>
 type MapeoDocTablesMap = {
@@ -87,12 +87,8 @@ export class DataType<
 
   getByDocId(
     docId: string,
-    opts?: { mustBeFound?: true; lang?: string }
+    opts?: { lang?: string }
   ): Promise<TDoc & { forks: string[] }>
-  getByDocId(
-    docId: string,
-    opts?: { mustBeFound?: boolean; lang?: string }
-  ): Promise<null | (TDoc & { forks: string[] })>
 
   getByVersionId(versionId: string, opts?: { lang?: string }): Promise<TDoc>
