+++ test/data-type/index.js
@@ -1,25 +1,26 @@
 import test from 'node:test'
 import assert from 'node:assert/strict'
-import { DataStore } from '../src/datastore/index.js'
+import { DataStore } from '../../src/datastore/index.js'
 import {
   createCoreManager,
   waitForCores,
   replicate,
-} from './helpers/core-manager.js'
+} from '../helpers/core-manager.js'
 import RAM from 'random-access-memory'
 import crypto from 'hypercore-crypto'
-import { observationTable, translationTable } from '../src/schema/project.js'
-import { DataType, kCreateWithDocId } from '../src/datatype/index.js'
-import { IndexWriter } from '../src/index-writer/index.js'
-import { NotFoundError } from '../src/errors.js'
+import { observationTable } from '../../src/schema/project.js'
+import { DataType, kCreateWithDocId } from '../../src/datatype/index.js'
+import { IndexWriter } from '../../src/index-writer/index.js'
+import { NotFoundError } from '../../src/errors.js'
 
 import Database from 'better-sqlite3'
 import { drizzle } from 'drizzle-orm/better-sqlite3'
 import { migrate } from 'drizzle-orm/better-sqlite3/migrator'
 import { randomBytes } from 'crypto'
-import TranslationApi from '../src/translation-api.js'
 import { getProperty } from 'dot-prop'
-import { decode, decodeBlockPrefix, parseVersionId } from '@comapeo/schema'
+import { parseVersionId } from '@comapeo/schema'
+
+import { testenv } from './test-helpers.js'
 
 /** @type {import('@comapeo/schema').ObservationValue} */
 const obsFixture = {
@@ -45,7 +46,8 @@ test('private createWithDocId() method', async () => {
   const sqlite = new Database(':memory:')
   const db = drizzle(sqlite)
   migrate(db, {
-    migrationsFolder: new URL('../drizzle/project', import.meta.url).pathname,
+    migrationsFolder: new URL('../../drizzle/project', import.meta.url)
+      .pathname,
   })
 
   const coreManager = createCoreManager()
@@ -81,7 +83,8 @@ test('private createWithDocId() method throws when doc exists', async () => {
   const sqlite = new Database(':memory:')
   const db = drizzle(sqlite)
   migrate(db, {
-    migrationsFolder: new URL('../drizzle/project', import.meta.url).pathname,
+    migrationsFolder: new URL('../../drizzle/project', import.meta.url)
+      .pathname,
   })
 
   const coreManager = createCoreManager()
@@ -209,14 +212,6 @@ test('getByDocId() throws if no document exists with that ID', async () => {
   await assert.rejects(() => dataType.getByDocId('foo bar'), NotFoundError)
 })
 
-test('getByDocId() can return null if no document exists with that ID', async () => {
-  const { dataType } = await testenv({ projectKey: randomBytes(32) })
-  assert.equal(
-    await dataType.getByDocId('foo bar', { mustBeFound: false }),
-    null
-  )
-})
-
 test('delete()', async () => {
   const projectKey = randomBytes(32)
   const { dataType } = await testenv({ projectKey })
@@ -311,82 +306,3 @@ test('translation', async () => {
     `not passing a a language code returns the untranslated message`
   )
 })
-
-/**
- * @param {object} [opts={}]
- * @param {Buffer} [opts.projectKey]
- */
-async function testenv(opts = {}) {
-  const sqlite = new Database(':memory:')
-  const db = drizzle(sqlite)
-  migrate(db, {
-    migrationsFolder: new URL('../drizzle/project', import.meta.url).pathname,
-  })
-
-  const coreManager = createCoreManager({ ...opts, db })
-
-  const indexWriter = new IndexWriter({
-    tables: [observationTable, translationTable],
-    sqlite,
-  })
-
-  const dataStore = new DataStore({
-    coreManager,
-    namespace: 'data',
-    batch: async (entries) => indexWriter.batch(entries),
-    storage: () => new RAM(),
-    reindex: false,
-  })
-
-  const configDataStore = new DataStore({
-    coreManager,
-    namespace: 'config',
-    batch: async (entries) => {
-      /** @type {import('multi-core-indexer').Entry[]} */
-      const entriesToIndex = []
-      for (const entry of entries) {
-        const { schemaName } = decodeBlockPrefix(entry.block)
-        try {
-          if (schemaName === 'translation') {
-            const doc = decode(entry.block, {
-              coreDiscoveryKey: entry.key,
-              index: entry.index,
-            })
-            assert(
-              doc.schemaName === 'translation',
-              'expected a translation doc'
-            )
-            translationApi.index(doc)
-          }
-          entriesToIndex.push(entry)
-        } catch {
-          // Ignore errors thrown by values that can't be decoded for now
-        }
-      }
-      const indexed = await indexWriter.batch(entriesToIndex)
-      return indexed
-    },
-    storage: () => new RAM(),
-    reindex: false,
-  })
-
-  const translationDataType = new DataType({
-    dataStore: configDataStore,
-    table: translationTable,
-    db,
-    getTranslations: () => {
-      throw new Error('Cannot get translations for translations')
-    },
-  })
-
-  const translationApi = new TranslationApi({ dataType: translationDataType })
-
-  const dataType = new DataType({
-    dataStore,
-    table: observationTable,
-    db,
-    getTranslations: translationApi.get.bind(translationApi),
-  })
-
-  return { coreManager, dataType, dataStore, translationApi }
-}
